Add configuration and delete the old lsp_lua.lua file
